Improved code after code review
